wificond: Assert failure for three continuous NODEV error
If (sched) scan failed with NODEV error, wificond considers
driver in bad state, and asserts failure to notify framework
to restart wifi stack.

This brings in trouble that if driver supports SSR, and when
it happens, wifi interface is removed, and it will cost some
time to recover. If framework tries to restart wifi stack during
this period, it will fail and wifi will be disabled.

To cover driver SSR case, change wificond to assert failure
for three continuous NODEV error.

CRs-Fixed: 2255918
Change-Id: Ief912b0f5145c587f019344aa9e964033d15494f